Fix NotConnected exception

+++ M13_Project_1_AKAE/src/main/java/cat/iticbcn/demo/controllers/RestAppController.java
@@ -14,6 +14,7 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
+import cat.iticbcn.demo.Exception.CompanyAndOfferNotConnectedException;
 import cat.iticbcn.demo.Exception.CompanyNotFoundException;
 import cat.iticbcn.demo.Exception.OfferNotFoundException;
 import cat.iticbcn.demo.bean.Company;
@@ -26,7 +27,7 @@ public class RestAppController {
 
 	@Autowired
 	private final OfferRepository offerRepository;
-	
+
 	@Autowired
 	private final CompanyRepository companyRepository;
 
@@ -34,58 +35,54 @@ public class RestAppController {
 		this.offerRepository = offerRepository;
 		this.companyRepository = companyRepository;
 	}
-	
-	//COMPANY METHODS----------------------------------------------
-	  // Aggregate root
-	  // tag::get-aggregate-root[]
-	  @GetMapping("/companies")
-	  List<Company> allCompanies() {
-	    return companyRepository.findAll();
-	  }
-	  // end::get-aggregate-root[]
 
-	  @PostMapping("/companies")
-	  Company newCompany(@RequestBody Company newCompany) {
-	    return companyRepository.save(newCompany);
-	  }
-	  
+	// COMPANY METHODS----------------------------------------------
+	// Aggregate root
+	// tag::get-aggregate-root[]
+	@GetMapping("/companies")
+	List<Company> allCompanies() {
+		return companyRepository.findAll();
+	}
+	// end::get-aggregate-root[]
 
-	  // Single item
-	  
-	  @GetMapping("/companies/{id}")
-	  Company oneCompany(@PathVariable Long id) {
-	    
-		  return companyRepository.findById(id)
-			      .orElseThrow(() -> new CompanyNotFoundException(id));
-	  }
-	  
-	  @PutMapping("/companies/{id}")
-	  Company replaceCompany(@RequestBody Company newCompany, @PathVariable Long id) {
-	    
-	    return companyRepository.findById(id)
-	      .map(company -> {
-	        company.setName(newCompany.getName());
-	        company.setEmployees(newCompany.getEmployees());
-	        company.setSocialSecurityNumber(newCompany.getSocialSecurityNumber());
-	        company.setOwner(newCompany.getOwner());
-	        company.setAddress(newCompany.getAddress());
-	        company.setPhoneNumber(newCompany.getPhoneNumber());
-	        company.setEmail(newCompany.getEmail());
-	        company.setType(newCompany.getType());
-	        return companyRepository.save(company);
-	      })
-	      .orElseGet(() -> {
-	        newCompany.setId(id);
-	        return companyRepository.save(newCompany);
-	      });
-	  }
+	@PostMapping("/companies")
+	Company newCompany(@RequestBody Company newCompany) {
+		return companyRepository.save(newCompany);
+	}
 
-	  @DeleteMapping("/companies/{id}")
-	  void deleteCompany(@PathVariable Long id) {
-	    companyRepository.deleteById(id);
-	  }
-	  
-	//OFFER METHODS----------------------------------------------
+	// Single item
+
+	@GetMapping("/companies/{id}")
+	Company oneCompany(@PathVariable Long id) {
+
+		return companyRepository.findById(id).orElseThrow(() -> new CompanyNotFoundException(id));
+	}
+
+	@PutMapping("/companies/{id}")
+	Company replaceCompany(@RequestBody Company newCompany, @PathVariable Long id) {
+
+		return companyRepository.findById(id).map(company -> {
+			company.setName(newCompany.getName());
+			company.setEmployees(newCompany.getEmployees());
+			company.setSocialSecurityNumber(newCompany.getSocialSecurityNumber());
+			company.setOwner(newCompany.getOwner());
+			company.setAddress(newCompany.getAddress());
+			company.setPhoneNumber(newCompany.getPhoneNumber());
+			company.setEmail(newCompany.getEmail());
+			company.setType(newCompany.getType());
+			return companyRepository.save(company);
+		}).orElseGet(() -> {
+			newCompany.setId(id);
+			return companyRepository.save(newCompany);
+		});
+	}
+
+	@DeleteMapping("/companies/{id}")
+	void deleteCompany(@PathVariable Long id) {
+		companyRepository.deleteById(id);
+	}
+
+	// OFFER METHODS----------------------------------------------
 	// Aggregate root
 	// tag::get-aggregate-root[]
 	@GetMapping("/offers")
@@ -96,22 +93,21 @@ public class RestAppController {
 
 	// Single item
 
-		@GetMapping("/offers/{id}")
-		Offer oneOffer(@PathVariable Long id) {
-			return offerRepository.findById(id).orElseThrow(() -> new OfferNotFoundException(id));
-		}
-		
-	@GetMapping(value="companies/{id}/offers")
-	List<Offer> getOffer(@PathVariable("id") Long id){
+	@GetMapping("/offers/{id}")
+	Offer oneOffer(@PathVariable Long id) {
+		return offerRepository.findById(id).orElseThrow(() -> new OfferNotFoundException(id));
+	}
+
+	@GetMapping(value = "companies/{id}/offers")
+	List<Offer> getOffer(@PathVariable("id") Long id) {
 		Optional<Company> company = companyRepository.findById(id);
 		List<Offer> offers = (List<Offer>) company.get().getOffers();
 		return offers;
 	}
-	
-		
-	@PostMapping(value="companies/{id}/offers")
-	public ResponseEntity<Offer> createOfferCompany(@RequestBody Offer offer, @PathVariable("id") Long id){
-		try{
+
+	@PostMapping(value = "companies/{id}/offers")
+	public ResponseEntity<Offer> createOfferCompany(@RequestBody Offer offer, @PathVariable("id") Long id) {
+		try {
 			Optional<Company> company = companyRepository.findById(id);
 			offer.setCompany(company.get());
 			Offer newof = offerRepository.save(offer);
@@ -119,7 +115,7 @@ public class RestAppController {
 		} catch (Exception e) {
 			throw new CompanyNotFoundException(id);
 		}
-		
+
 	}
 
 	@PutMapping("/offers/{id}")
@@ -135,10 +131,18 @@ public class RestAppController {
 	@DeleteMapping("companies/{idCo}/offers/{idOf}")
 	@Transactional
 	void deleteOffer(@PathVariable Long idCo, @PathVariable Long idOf) {
-		Optional<Company> company = Optional.of(companyRepository.findById(idCo).orElseThrow(() -> new CompanyNotFoundException(idCo)));
-		Optional<Offer> offer = Optional.of(offerRepository.findById(idOf).orElseThrow(() -> new OfferNotFoundException(idOf)));
+
+		Optional<Company> company = Optional
+				.of(companyRepository.findById(idCo).orElseThrow(() -> new CompanyNotFoundException(idCo)));
+		Optional<Offer> offer = Optional
+				.of(offerRepository.findById(idOf).orElseThrow(() -> new OfferNotFoundException(idOf)));
+		if (!company.get().getOffers().contains(offer.get())) {
+
+			throw new CompanyAndOfferNotConnectedException(idCo, idOf);
+
+		}
 		company.get().getOffers().remove(offer.get());
 		offerRepository.deleteById(idOf);
-		
+
 	}
 }
